# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solana zkAPI USDC + Proxy",
    "version": "1.0.0-design",
    "description": "Implementation contract, not a deployed service. JSON amounts are decimal strings. Public control objects reject unknown fields; provider nested payloads require adapter validation."
  },
  "servers": [
    {
      "url": "https://zkapi.example.invalid"
    }
  ],
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "operationId": "chatCompletions",
        "security": [
          {
            "ProxyToken": []
          },
          {
            "AnthropicProxyKey": []
          }
        ],
        "description": "Proxy only; validate provider-native nested fields against pinned adapter schema. Reject unsupported modalities/tools before dispatch. Idempotency retry returns status, never reruns inference.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider-native JSON or SSE (count_tokens: JSON only).",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                },
                "description": "Relative control operation status URL."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Provider-compatible error; X-Zkapi-Error-Code header. No raw provider secrets."
          },
          "409": {
            "description": "Operation in progress, response not replayable, or conflicting identity. Never redispatch.",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Zkapi-Error-Code": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/responses": {
      "post": {
        "operationId": "responses",
        "security": [
          {
            "ProxyToken": []
          },
          {
            "AnthropicProxyKey": []
          }
        ],
        "description": "Proxy only; validate provider-native nested fields against pinned adapter schema. Reject unsupported modalities/tools before dispatch. Idempotency retry returns status, never reruns inference.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponsesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider-native JSON or SSE (count_tokens: JSON only).",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                },
                "description": "Relative control operation status URL."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Provider-compatible error; X-Zkapi-Error-Code header. No raw provider secrets."
          },
          "409": {
            "description": "Operation in progress, response not replayable, or conflicting identity. Never redispatch.",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Zkapi-Error-Code": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/messages": {
      "post": {
        "operationId": "messages",
        "security": [
          {
            "ProxyToken": []
          },
          {
            "AnthropicProxyKey": []
          }
        ],
        "description": "Proxy only; validate provider-native nested fields against pinned adapter schema. Reject unsupported modalities/tools before dispatch. Idempotency retry returns status, never reruns inference.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          {
            "name": "anthropic-version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider-native JSON or SSE (count_tokens: JSON only).",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                },
                "description": "Relative control operation status URL."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Provider-compatible error; X-Zkapi-Error-Code header. No raw provider secrets."
          },
          "409": {
            "description": "Operation in progress, response not replayable, or conflicting identity. Never redispatch.",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Zkapi-Error-Code": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/messages/count_tokens": {
      "post": {
        "operationId": "countTokens",
        "security": [
          {
            "ProxyToken": []
          },
          {
            "AnthropicProxyKey": []
          }
        ],
        "description": "Proxy only; validate provider-native nested fields against pinned adapter schema. Reject unsupported modalities/tools before dispatch. Idempotency retry returns status, never reruns inference.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          {
            "name": "anthropic-version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountTokensRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider-native JSON or SSE (count_tokens: JSON only).",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                },
                "description": "Relative control operation status URL."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Provider-compatible error; X-Zkapi-Error-Code header. No raw provider secrets."
          },
          "409": {
            "description": "Operation in progress, response not replayable, or conflicting identity. Never redispatch.",
            "headers": {
              "X-Zkapi-Operation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "X-Zkapi-Status-Url": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Zkapi-Error-Code": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "operationId": "models",
        "security": [
          {
            "ProxyToken": []
          },
          {
            "AnthropicProxyKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider-compatible allowlisted model list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": [
                    "system",
                    "developer",
                    "user",
                    "assistant",
                    "tool"
                  ]
                },
                "content": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "name": {
                  "type": "string"
                },
                "tool_call_id": {
                  "type": "string"
                },
                "tool_calls": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "const": "function"
                      },
                      "function": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "arguments": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "name",
                          "arguments"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "id",
                      "type",
                      "function"
                    ],
                    "additionalProperties": false
                  }
                }
              },
              "required": [
                "role"
              ],
              "additionalProperties": false
            },
            "minItems": 1
          },
          "max_completion_tokens": {
            "type": "integer",
            "minimum": 1
          },
          "max_tokens": {
            "type": "integer",
            "minimum": 1
          },
          "stream": {
            "type": "boolean"
          },
          "temperature": {
            "type": "number"
          },
          "top_p": {
            "type": "number"
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "const": "function"
                },
                "function": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "parameters": {
                      "type": "object",
                      "properties": {},
                      "required": [],
                      "additionalProperties": true
                    },
                    "strict": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "name",
                    "parameters"
                  ],
                  "additionalProperties": false
                }
              },
              "required": [
                "type",
                "function"
              ],
              "additionalProperties": false
            }
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": true
              }
            ]
          },
          "stream_options": {
            "type": "object",
            "properties": {
              "include_usage": {
                "type": "boolean"
              }
            },
            "required": [],
            "additionalProperties": false
          }
        },
        "required": [
          "model",
          "messages"
        ],
        "additionalProperties": false,
        "not": {
          "required": [
            "max_tokens",
            "max_completion_tokens"
          ]
        }
      },
      "ResponsesRequest": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              }
            ]
          },
          "instructions": {
            "type": "string"
          },
          "max_output_tokens": {
            "type": "integer",
            "minimum": 1
          },
          "stream": {
            "type": "boolean"
          },
          "store": {
            "type": "boolean",
            "const": false,
            "default": false
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "required": [],
              "additionalProperties": true
            }
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": true
              }
            ]
          }
        },
        "required": [
          "model",
          "input",
          "max_output_tokens"
        ],
        "additionalProperties": false
      },
      "MessagesRequest": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": [
                    "user",
                    "assistant"
                  ]
                },
                "content": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {},
                        "required": [],
                        "additionalProperties": true
                      }
                    }
                  ]
                }
              },
              "required": [
                "role",
                "content"
              ],
              "additionalProperties": false
            },
            "minItems": 1
          },
          "max_tokens": {
            "type": "integer",
            "minimum": 1
          },
          "system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              }
            ]
          },
          "stream": {
            "type": "boolean"
          },
          "temperature": {
            "type": "number"
          },
          "top_p": {
            "type": "number"
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "required": [],
              "additionalProperties": true
            }
          },
          "tool_choice": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": true
          }
        },
        "required": [
          "model",
          "messages",
          "max_tokens"
        ],
        "additionalProperties": false
      },
      "CountTokensRequest": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": [
                    "user",
                    "assistant"
                  ]
                },
                "content": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {},
                        "required": [],
                        "additionalProperties": true
                      }
                    }
                  ]
                }
              },
              "required": [
                "role",
                "content"
              ],
              "additionalProperties": false
            },
            "minItems": 1
          },
          "system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {},
                  "required": [],
                  "additionalProperties": true
                }
              }
            ]
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "required": [],
              "additionalProperties": true
            }
          },
          "tool_choice": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": true
          }
        },
        "required": [
          "model",
          "messages"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "ControlToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "zkc1.<uuid>.<32-byte-secret-base64url>"
      },
      "ProxyToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "zkp1.<uuid>.<32-byte-secret-base64url>"
      },
      "AdminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin-only credential on private listener; not a user token."
      },
      "AnthropicProxyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Same proxy token, never an upstream API key."
      }
    }
  }
}